backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
pydantic
requests
PyJWT>=2.8,<3